+++ CHANGELOG.md
@@ -71,3 +71,4 @@ Upcoming changes compared to [NetCoreDbg](https://github.com/Samsung/netcoredbg)
 - Fixed some methods `void *&` (`PVOID &`) parameters to `void **`.
 - Fixed stacktrace for exception in async methods (exception rethrow with `System.Runtime.ExceptionServices.ExceptionDispatchInfo.Throw()`).
 - Fixed exception type name fail handling logic in GetExceptionDetails().
+- Fixed disable JIT optimization related logic.